Created Anim Instance for sentinel rebel.
SentinelRebelAnimInstance.h and .cpp has the base functionality for running the belica animation(in my case the player sentinel rebel).

The player(basically the rebel component access is retrived by using the TryGetPawnOwner method).
isAccelerating is the check for movement of the player - It is acquired by the GetCurrentAcceleration vector's size value(basically size is vector's magnitude and when the size is 0 we are essentially saying the player isn't moving.).

Player anim blue print is created and derived from anim instance(SentinelRebelAnimInstance)
The player anim blueprint is referenced in the Anim class under animation in mesh component of the player.
The UpdateAnimationProperities is called from the event graph of the new created player anim blueprint class.