Updated the line tracing for bullet impact.
- Used two line traces.
- First trace from crosshair outward - 50,000 units.(if something hits then second trace end/target point will be that hit point).
- Second trace is from muzzle to the hit/end/target point from the first trace to make sure if there is any object in the path of the bullet(travelling from muzzle to the hit point).

- Both the traces are imp.

First trace is not the actual bullet travel path, it's only made in order to get the first blocking hit from the crosshair, so that our aiming can be precise. Then with this hit location we run the 2nd trace to make sure the bullet will actually get there (to avoid bullets running through a cover wall or corner for example) and then we make the effects to show  the path from the barrel to the final hit location.

It's a bit tricky to understand but it's a good solution to aim from the camera but still fire from the 3rd person character.